chore: merged models and models-info command
